requirements.txt
transformers
openai==1.93.3
matplotlib
numpy
jupyter
python-dotenv==1.1.1